requirements.txt
streamlit==0.71.0
pandas==1.1.3
base58==2.0.1
numpy
scikit-learn
Pillow

